requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
requests
pymupdf
langchain
langchain-groq          
langchain-google-genai  
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv